+++ src/app/modules/angular-slickgrid/editors/index.ts
@@ -4,10 +4,10 @@ export * from './dateEditor';
 export * from './dualInputEditor';
 export * from './editors.index';
 export * from './floatEditor';
+export * from './inputEditor';
 export * from './integerEditor';
 export * from './longTextEditor';
 export * from './multipleSelectEditor';
 export * from './selectEditor';
 export * from './singleSelectEditor';
 export * from './sliderEditor';
-export * from './textEditor';
